src/communication/mqtt_client.py
"""
MQTT Client
MQTT client cho real-time data transmission và remote monitoring
"""

from typing import Dict, Any, Optional, Callable, List
import logging
import json
import ssl
from paho.mqtt.client import Client as MQTTClient
from threading import Lock


class MQTTClient:
    """
    MQTT client cho IoT Health Monitoring System
    
    Attributes:
        config (Dict): MQTT configuration
        client (MQTTClient): Paho MQTT client instance
        is_connected (bool): Connection status
        topic_prefix (str): Topic prefix for all messages
        message_handlers (Dict): Dictionary of message handlers
        connection_lock (Lock): Thread lock for connection operations
        retry_count (int): Current retry count for reconnection
        last_will_topic (str): Last will topic
    """
    
    def __init__(self, config: Dict[str, Any]):
        """
        Initialize MQTT client
        
        Args:
            config: MQTT configuration dictionary
        """
        pass
    
    def connect(self) -> bool:
        """
        Connect to MQTT broker
        
        Returns:
            bool: True if connection successful
        """
        pass
    
    def disconnect(self) -> bool:
        """
        Disconnect from MQTT broker
        
        Returns:
            bool: True if disconnection successful
        """
        pass
    
    def publish_vital_signs(self, patient_id: str, vital_data: Dict[str, Any]) -> bool:
        """
        Publish vital signs data
        
        Args:
            patient_id: Patient identifier
            vital_data: Dictionary containing vital signs data
            
        Returns:
            bool: True if publish successful
        """
        pass
    
    def publish_blood_pressure_features(self, patient_id: str, bp_features: Dict[str, Any]) -> bool:
        """
        Publish blood pressure features
        
        Args:
            patient_id: Patient identifier
            bp_features: Blood pressure oscillation features
            
        Returns:
            bool: True if publish successful
        """
        pass
    
    def publish_alert(self, patient_id: str, alert_data: Dict[str, Any]) -> bool:
        """
        Publish alert notification
        
        Args:
            patient_id: Patient identifier
            alert_data: Alert information
            
        Returns:
            bool: True if publish successful
        """
        pass
    
    def subscribe_to_predictions(self, patient_id: str, callback: Callable) -> bool:
        """
        Subscribe to AI prediction results
        
        Args:
            patient_id: Patient identifier
            callback: Function to handle prediction messages
            
        Returns:
            bool: True if subscription successful
        """
        pass
    
    def subscribe_to_commands(self, patient_id: str, callback: Callable) -> bool:
        """
        Subscribe to remote commands
        
        Args:
            patient_id: Patient identifier
            callback: Function to handle command messages
            
        Returns:
            bool: True if subscription successful
        """
        pass
    
    def _setup_ssl_context(self) -> ssl.SSLContext:
        """
        Setup SSL context for secure connection
        
        Returns:
            SSL context for MQTT connection
        """
        pass
    
    def _on_connect(self, client, userdata, flags, rc):
        """
        Callback for MQTT connection event
        
        Args:
            client: MQTT client instance
            userdata: User data
            flags: Connection flags
            rc: Return code
        """
        pass
    
    def _on_disconnect(self, client, userdata, rc):
        """
        Callback for MQTT disconnection event
        
        Args:
            client: MQTT client instance
            userdata: User data
            rc: Return code
        """
        pass
    
    def _on_message(self, client, userdata, msg):
        """
        Callback for incoming MQTT messages
        
        Args:
            client: MQTT client instance
            userdata: User data
            msg: Received message
        """
        pass
    
    def _on_publish(self, client, userdata, mid):
        """
        Callback for message publish confirmation
        
        Args:
            client: MQTT client instance
            userdata: User data
            mid: Message ID
        """
        pass
    
    def _on_subscribe(self, client, userdata, mid, granted_qos):
        """
        Callback for subscription confirmation
        
        Args:
            client: MQTT client instance
            userdata: User data
            mid: Message ID
            granted_qos: Granted QoS levels
        """
        pass
    
    def _build_topic(self, *topic_parts) -> str:
        """
        Build topic string with prefix
        
        Args:
            topic_parts: Topic parts to join
            
        Returns:
            Complete topic string
        """
        pass
    
    def _handle_connection_error(self, error: Exception):
        """
        Handle connection errors and retry logic
        
        Args:
            error: Connection error exception
        """
        pass
    
    def _reconnect_with_backoff(self):
        """
        Reconnect with exponential backoff
        """
        pass
    
    def add_message_handler(self, topic_pattern: str, handler: Callable):
        """
        Add message handler for specific topic pattern
        
        Args:
            topic_pattern: Topic pattern to match
            handler: Function to handle messages
        """
        pass
    
    def remove_message_handler(self, topic_pattern: str):
        """
        Remove message handler for topic pattern
        
        Args:
            topic_pattern: Topic pattern to remove
        """
        pass
    
    def get_connection_status(self) -> Dict[str, Any]:
        """
        Get current connection status and statistics
        
        Returns:
            Dictionary containing connection information
        """
        pass
    
    def set_last_will(self, topic: str, message: str, qos: int = 1, retain: bool = True):
        """
        Set last will and testament message
        
        Args:
            topic: Last will topic
            message: Last will message
            qos: Quality of service level
            retain: Whether to retain message
        """
        pass