requirements.txt
# IoT Health Monitoring System - Requirements

# Core libraries
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# GUI framework
kivy>=2.1.0
kivymd>=1.0.0

# Sensor libraries
adafruit-circuitpython-max30x>=3.2.0
adafruit-circuitpython-ads1x15>=2.2.0
w1thermsensor>=2.0.0

# Communication
paho-mqtt>=1.6.0
requests>=2.28.0
websockets>=10.0

# Database
sqlite3  # Built-in
sqlalchemy>=1.4.0

# AI/ML
scikit-learn>=1.0.0
tensorflow-lite>=2.8.0

# Utilities
PyYAML>=6.0
python-dotenv>=0.19.0
schedule>=1.1.0

# Audio/Alerts
pygame>=2.1.0

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Hardware interface
RPi.GPIO>=0.7.1
spidev>=3.5
smbus2>=0.4.0